Move initialvalue as a config file
