+++ apps/console/src/features/feature-gate/reducer/feature-gate.tsx
@@ -18,8 +18,8 @@
 import React, { useReducer } from "react";
 import { FeatureGateAction, FeatureGateActionTypes } from "../actions/feature-gate";
 import { FeatureGateContext } from "../context/feature-gate";
-import { FeatureGateInterface, featureGateConfig } from "../models/feature-gate";
-
+import { FeatureGateInterface } from "../models/feature-gate";
+import { featureGateConfig } from "../configs/feature-gate";
 
 export const featureGateReducer = (
     state: FeatureGateInterface,
